refactor: 인증 및 보안 관련 클래스 패키지 구조 재조정
- **패키지 변경**:
  - `config` 패키지 내부 클래스를 `config.auth` 및 `config.web`으로 분리
  - 인증 관련 클래스(JWT, 필터 등): `config.auth`
  - 웹 설정 클래스(CORS 등): `config.web`
- **파일 수정**:
  - 패키지 이동에 따른 import 경로 업데이트
- 패키지 구조 개선으로 클래스 역할 및 관리 용이성 향상
