+++ src/main/java/com/tropical/backend/auth/controller/AuthController.java
@@ -8,8 +8,8 @@ import com.tropical.backend.auth.dto.response.UserResponse;
 import com.tropical.backend.auth.entity.User;
 import com.tropical.backend.auth.service.UserConsentService;
 import com.tropical.backend.auth.service.UserService;
-import com.tropical.backend.config.JwtAuthenticationFilter;
-import com.tropical.backend.config.JwtTokenProvider;
+import com.tropical.backend.config.auth.JwtAuthenticationFilter;
+import com.tropical.backend.config.auth.JwtTokenProvider;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
