+++ src/main/java/com/tropical/backend/config/security/SecurityConfig.java
@@ -1,6 +1,6 @@
-package com.tropical.backend.config;
+package com.tropical.backend.config.security;
 
-import com.tropical.backend.config.security.CustomAccessDeniedHandler;
+import com.tropical.backend.config.auth.JwtAuthenticationFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
