+++ src/main/java/com/tropical/backend/config/web/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.tropical.backend.config;
+package com.tropical.backend.config.web;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
